Restructured builtins into individual modules
