+++ src/rsh/builtins/echo.rs
@@ -1,0 +1,15 @@
+use std::path::PathBuf;
+
+use rsh::State;
+
+pub fn echo(s: &mut State) -> i32 {
+    if s.argv[1] == "-n" {
+        let strings = &s.argv[2..s.argv.len()].join(" ");
+        print!("{}", strings);
+    } else {
+        let strings = &s.argv[1..s.argv.len()].join(" ");
+        println!("{}", strings);
+    }
+
+    0
+}
